Refactor: Move generic helpers from simulation to utils
This refactoring moves generic helper functions from `clintrials/simulation.py` to a new `clintrials/utils.py` module. This improves code organization, promotes reusability, and helps to avoid potential circular dependencies.

The following functions were moved:
- `_open_json_local`
- `_open_json_url`
- `go_fetch_json_sims` (renamed to `fetch_json_from_files`)
- `filter_sims` (renamed to `filter_list_of_dicts`)
- `map_reduce_files`
- `invoke_map_reduce_function_map` (renamed to `invoke_map_reduce_on_list`)
- `reduce_maps_by_summing`
- `multiindex_dataframe_from_tuple_map`

Deprecation shims have been added to `clintrials/simulation.py` for all moved functions to ensure backward compatibility. All internal references to these functions have been updated to use the new `clintrials.utils` module.
